Added some epic styling and solved some behavioural search issues
Added styling for most of the home view. Added a grid-like panel for the search-results, while also adding a thumbnail.

Fixed some issues like not reloading results when deleting, not unselecting facets when writing something new in the search box.
Added a new behaviour so that when deleting the whole text, it will show 0 results